+++ paying_for_college/views.py
@@ -47,8 +47,10 @@ else:  # pragma: no cover
 URL_ROOT = 'paying-for-college2'
 EXPENSE_FILE = '{}/fixtures/bls_data.json'.format(BASEDIR)
 NO_SCHOOL_ERROR = "No active school could be found for iped ID {0}"
-ID_ERROR = "Illegal offer ID; only characters 0-9 and a-f are allowed."
-PID_ERROR = "Error: Illegal characters in program code '{}'"
+OID_WARNING = "Error: Illegal offer ID"
+# only characters 0-9 and a-f are allowed in offer IDs
+PID_WARNING = "Error: Illegal program code"
+# program code isn't 40 chars or contains Illegal characters: ; < > { }
 
 
 def get_json_file(filename):
@@ -61,14 +63,17 @@ def get_json_file(filename):
 
 def validate_oid(oid):
     """
-    make sure an oid contains only hex values 0-9 a-f A-F
+    make sure an oid contains only hex values 0-9 a-f A-F and is 40 characters
     return True if the oid is valid
     """
     find_illegal = re.search('[^0-9a-fA-F]+', oid)
     if find_illegal:
         return False
     else:
-        return True
+        if len(oid) == 40:
+            return True
+        else:
+            return False
 
 
 def validate_pid(pid):
@@ -100,9 +105,12 @@ def get_school(schoolID):
     try:
         school = School.objects.get(school_id=int(schoolID))
     except:
-        return ''
+        return None
     else:
-        return school
+        if school.operating is False:
+            return None
+        else:
+            return school
 
 
 def get_program(school, programCode):
@@ -126,44 +134,60 @@ class BaseTemplateView(TemplateView):
         return context
 
 
-class OfferView(TemplateView):  # TODO log errors
+class OfferView(TemplateView):
     """consult values in querystring and deliver school/program data"""
 
     def get(self, request):
+        school = None
+        program = None
+        program_data = 'null'
+        school_data = 'null'
+        warning = ''
+        OID = ''
+        if 'oid' in request.GET and request.GET['oid']:
+            OID = request.GET['oid']
+        else:
+            warning = "Warning: URL doesn't contain an offer ID"
+        if OID and validate_oid(OID) is False:
+            warning = OID_WARNING
+            OID = ''
         if 'iped' in request.GET and request.GET['iped']:
             iped = request.GET['iped']
             school = get_school(iped)
-            if not school or not school.operating:
-                return HttpResponseBadRequest(NO_SCHOOL_ERROR.format(iped))
-            if 'oid' in request.GET:
-                OID = request.GET['oid']
+            if school:
+                school_data = school.as_json()
+                if 'pid' in request.GET and request.GET['pid']:
+                    PID = request.GET['pid']
+                    if not validate_pid(PID):
+                        warning = PID_WARNING
+                        PID = ''
+                    if PID:
+                        programs = Program.objects.filter(program_code=PID,
+                                                          institution=school).order_by('-pk')
+                        if programs:
+                            program = programs[0]
+                            program_data = program.as_json()
+                        else:
+                            warning = ("Warning: No program could be found "
+                                       "for program_code '{}'".format(PID))
+                else:
+                    warning = "Warning: URL doesn't contain a program code"
             else:
-                OID = ''
-            if OID and validate_oid(OID) is False:
-                return HttpResponseBadRequest(ID_ERROR)
-            program_data = json.dumps({})
-            program = ''
-            if 'pid' in request.GET and request.GET['pid']:
-                PID = request.GET['pid']
-                if not validate_pid(PID):
-                    return HttpResponseBadRequest(PID_ERROR.format(PID))
-                programs = Program.objects.filter(program_code=PID,
-                                                  institution=school).order_by('-pk')
-                if programs:
-                    program = programs[0]
-                    program_data = program.as_json()
-            return render_to_response('worksheet.html',
-                                      {'data_js': "0",
-                                       'school': school,
-                                       'schoolData': school.as_json(),
-                                       'program': program,
-                                       'programData': program_data,
-                                       'oid': OID,
-                                       'base_template': BASE_TEMPLATE,
-                                       'url_root': URL_ROOT},
-                                      context_instance=RequestContext(request))
+                warning = ("Warning: No active school found "
+                           "for ID {}".format(iped))
         else:
-            return HttpResponseBadRequest("URL doesn't contain a school ID")
+                warning = "Warning: URL doesn't contain a school ID"
+        return render_to_response('worksheet.html',
+                                  {'data_js': "0",
+                                   'school': school,
+                                   'schoolData': school_data,
+                                   'program': program,
+                                   'programData': program_data,
+                                   'oid': OID,
+                                   'base_template': BASE_TEMPLATE,
+                                   'warning': warning,
+                                   'url_root': URL_ROOT},
+                                  context_instance=RequestContext(request))
 
 
 class LandingView(TemplateView):
@@ -242,7 +266,7 @@ class ProgramRepresentation(View):
             return HttpResponseBadRequest(format_error)
         PID = ids[1]
         if not validate_pid(PID):
-            return HttpResponseBadRequest(PID_ERROR.format(PID))
+            return HttpResponseBadRequest(PID_WARNING)
         program = self.get_program(program_code)
         if not program:
             p_error = ("Error: No program was found "