+++ CHANGELOG.md
@@ -17,7 +17,9 @@ We follow the [Semantic Versioning 2.0.0](http://semver.org/) format.
 - Update debt summary text for program durations less than one year.
 - Added privateLoanFee constant to the database
 - Added unicode-aware csv module
+- Refactored the offer view to allow bad school, program and offer IDs
 - Set total program debt multiplier minimum to 1.
+- Refactored the load_programs script to handle non-utf-8 encoding
 
 ## 2.1.4
 - Added string checking for program codes